Substitute envvars in WarningsChecker.substitute_envvar
